Added support for xlsx

+++ ulearn_scoring/_models.py
@@ -1,6 +1,7 @@
+from decimal import Decimal
 from typing import NamedTuple
 
 
 class StudentScores(NamedTuple):
     student: str
-    scores: int
+    scores: Decimal
